requirements.txt
flask
botbuilder-core
azure-ai-projects==1.0.0b10
azure-identity
gunicorn